[IMP] web: Convert Dropdown tests to hoot
This commit converts all legacy QUnit Dropdown to the new Hoot test framework.

- All test linked to dropdowns are under a dropdown folder.
- Test purely linked to the DropdownItem component are now in a seperated
file, same for DropdownGroup.
