+++ addons/web/static/tests/core/dropdown/dropdown_accordion_item.test.js
